Split parameters out of ForestSim class.
